# public/swagger.json
{
	"components": {
...
		"schemas": {
			"DefaultReponseBody": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RoleModel.Role": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FetchRolesResponseBody": {
				"properties": {
					"roles": {
						"items": {
							"$ref": "#/components/schemas/RoleModel.Role"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object"
			},
			"CreateRoleResponseBody": {
				"properties": {
					"role": {
						"$ref": "#/components/schemas/RoleModel.Role"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object"
			},
			"CreateRoleRequestBody": {
				"properties": {
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"name"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "university-research-website-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/role": {
			"get": {
				"operationId": "GetRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FetchRolesResponseBody"
								}
							}
						}
					},
					"400": {
						"description": "Invalid Input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DefaultReponseBody"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Invalid Input",
											"status": 400
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DefaultReponseBody"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Internal Server Error",
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateRoleResponseBody"
								}
							}
						}
					},
					"400": {
						"description": "Invalid Input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DefaultReponseBody"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Invalid Input",
											"status": 400
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DefaultReponseBody"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Internal Server Error",
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoleRequestBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}